Removed redundant code (point to circle)
