+++ chapters/004_GameDev_Basics/002_CollisionDetection/002_Narrow_Phase/003_Point_Circle.md
@@ -27,9 +27,4 @@ An example could be the following:
 ```{src='collisiondetection/point_circle' caption='Point to circle collision detection'}
 ```
 
-Again, the lazier version:
-
-```{src='collisiondetection/point_circle_lazy' caption='Shorter version of a point to circle collision detection'}
-```
-
 Although slightly more heavy, computation-wise, this algorithm still runs in O(1).
