--- dynamic_listings/javascript/default/collisiondetection/point_circle_lazy.js
@@ -1,16 +1,0 @@
-class Circle{
-    // Let's define a circle class/structure
-    constructor(){
-        this.center = new Point();
-        this.radius = 0;
-    }
-}
-
-function distance(A, B){
-    // Calculates the distance between two points
-    return Math.sqrt((A.x - B.x) ** 2 + (A.y - B.y) ** 2);
-}
-
-function circle_point_collision(A, B){
-    return distance(A.center, B) <= A.radius;
-}
