--- dynamic_listings/lua/default/collisiondetection/point_circle_lazy.lua
@@ -1,19 +1,0 @@
--- Let's define a circle class/structure
-Circle = {}
-
-function Circle:new(o)
-    -- This is an example constructor
-    o = o or {center=Point:new(), radius=0}
-    setmetatable(o, self)
-    self.__index = self
-    return o
-end
-
-local function distance(A, B)
-    -- Calculates the distance between two points
-    return math.sqrt((A.x - B.x)^2 + (A.y - B.y)^2)
-end
-
-local function circle_point_collision(A, B)
-    return distance(A.center, B) <= A.radius
-end
