+++ dynamic_listings/pseudocode/default/collisiondetection/point_circle.txt
@@ -4,15 +4,11 @@ structure Circle{
     int radius;
 }
 
-function distance(Point A, Point B) -> float{
+function distance(Point A, Point B) -> bool{
     // Calculates the distance between two points
     return square_root((A.x - B.x)^2 + (A.y - B.y)^2);
 }
 
 function circle_point_collision(Circle A, Point B) -> bool{
-    if (distance(A.center, B) <= A.radius){
-        return True;
-    }else{
-        return False;
-    }
+    return distance(A.center, B) <= A.radius;
 }
