--- dynamic_listings/pseudocode/default/collisiondetection/point_circle_lazy.txt
@@ -1,14 +1,0 @@
-structure Circle{
-    // Let's define a circle class/structure
-    Point center;
-    int radius;
-}
-
-function distance(Point A, Point B) -> bool{
-    // Calculates the distance between two points
-    return square_root((A.x - B.x)^2 + (A.y - B.y)^2);
-}
-
-function circle_point_collision(Circle A, Point B) -> bool{
-    return distance(A.center, B) <= A.radius;
-}
